requirements.txt
Jinja2 >= 2.11.2
Pillow >= 8.0.1
python-dateutil >= 2.8.1
